requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
backoff==2.2.1
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
choreographer==1.2.0
click==8.3.0
distro==1.9.0
flasgger==0.9.7.1
Flask==3.1.2
flask-sock==0.7.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaleido==1.1.0
logistro==2.0.0
MarkupSafe==3.0.3
mem0ai==1.0.0
mistune==3.1.4
narwhals==2.10.0
numpy==2.3.4
openai==2.6.1
orjson==3.11.4
packaging==25.0
pandas==2.3.3
plotly==6.3.1
pluggy==1.6.0
portalocker==3.2.0
posthog==6.7.11
protobuf==5.29.5
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pytest==8.4.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.15.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
simple-websocket==1.1.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
tabulate==0.9.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vanna==0.7.9
Werkzeug==3.1.3
wsproto==1.2.0
